refactor(workersvc): move no-receiver pgtype/error helpers to pgparams.go
Pure same-package code motion (PRD #921 M1, seam 1). Moves 15 no-receiver
helpers (pgText, stopReasonParam, textParam, sanitizeFailureReason,
clampWirePreservedPatch, stripNULParam, int8Param, pgIntPtr, pgFloat4Ptr,
pgUUID, pgTime, textPtr, coalesceInt, isUniqueViolation, uniqueViolationOn)
out of service.go into the new pgparams.go, byte-identically. The two
interspersed const blocks (maxFailureReasonRunes, PreservedPatchMaxBytes)
stay in service.go as they are shared across files. No exported-API change,
no behavior change; go doc output unchanged.
